AutoDiff_emma.py for tweaking
I only changed the __init__, specifically line 36 and 38, which deal with self.der and self.jacobian. The all() function did not work for me, and there are cases where the der_value needs to equal None. I'm not sure if this is completely right, but it works with the trig functions and sqrt function in my code.